requirements.txt
Django==4.0.1
python-dotenv==0.19.2
psycopg2-binary==2.8.5
djangorestframework==3.13.1
django_filter==21.1
drf-spectacular==0.22.1
djoser
social-auth-app-django==5.0.0
django-cors-headers==3.8.0

